chore(project): rename etherscan to history storage, add new etherscan storage
