+++ KyberNetwork/KyberNetwork/Screens/Tabbars/Earn/Controllers/EarnMenuViewController.swift
@@ -124,7 +124,7 @@ class EarnMenuViewController: KNBaseViewController {
     guard self.isViewLoaded else {
       return
     }
-    let pendingTransaction = EtherscanTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
+    let pendingTransaction = HistoryTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
       transaction.state == .pending
     }
     self.pendingTxIndicatorView.isHidden = pendingTransaction == nil
