+++ KyberNetwork/KyberNetwork/Screens/Tabbars/Earn/Controllers/EarnSwapViewController.swift
@@ -1131,7 +1131,7 @@ class EarnSwapViewController: KNBaseViewController, AbstractEarnViewControler {
     guard self.isViewLoaded else {
       return
     }
-    let pendingTransaction = EtherscanTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
+    let pendingTransaction = HistoryTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
       transaction.state == .pending
     }
     self.pendingTxIndicatorView.isHidden = pendingTransaction == nil
@@ -1190,11 +1190,11 @@ class EarnSwapViewController: KNBaseViewController, AbstractEarnViewControler {
     guard let token = self.viewModel.approvingToken else {
       return
     }
-    if EtherscanTransactionStorage.shared.getInternalHistoryTransaction().isEmpty {
+    if HistoryTransactionStorage.shared.getInternalHistoryTransaction().isEmpty {
       self.updateUIForSendApprove(isShowApproveButton: false)
       self.viewModel.approvingToken = nil
     }
-    let pending = EtherscanTransactionStorage.shared.getInternalHistoryTransaction().filter({ (item) -> Bool in
+    let pending = HistoryTransactionStorage.shared.getInternalHistoryTransaction().filter({ (item) -> Bool in
       return item.transactionDetailDescription.lowercased() == token.address.lowercased() && item.type == .allowance
     })
     if pending.isEmpty {
