+++ KyberNetwork/KyberNetwork/Screens/Transaction/Send/Controller/KSendTokenViewController.swift
@@ -405,7 +405,7 @@ extension KSendTokenViewController {
     guard self.isViewLoaded else {
       return
     }
-    let pendingTransaction = EtherscanTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
+    let pendingTransaction = HistoryTransactionStorage.shared.getInternalHistoryTransaction().first { transaction in
       transaction.state == .pending
     }
     self.pendingTxIndicatorView.isHidden = pendingTransaction == nil
